Mass commit. Camera, mesh adder, viewport, gizmo stuff (i'm sorry)

+++ RR_Godot/src/Core/Gui/ImportWindow.cs
@@ -5,7 +5,7 @@ using RR_Godot.Core;
 public class ImportWindow : FileDialog
 {
     private Global GlobalSettings;
-    
+
     // Called when the node enters the scene tree for the first time.
     public override void _Ready()
     {
@@ -18,9 +18,8 @@ public class ImportWindow : FileDialog
 
     public void OnFileSelected(string path)
     {
-        
         File selectedFile = new File();
-        selectedFile.Open(path, File.ModeFlags.Read);
+        selectedFile.Open(path, (int) File.ModeFlags.Read);
         string absolutePath = selectedFile.GetPathAbsolute();
         selectedFile.Close();
         GlobalSettings.ImportFile(absolutePath);
